rename next-auth/expo files
