+++ packages/next-auth/src/expo/index.ts
@@ -1,1 +1,1 @@
-export * from "./Auth";
+export * from "./expo"
